Extract out the various sample package descriptions as file-based fixtures
