+++ Tests/SwiftDependencyAuditTests/Fixtures/CustomPathPackage.swift
@@ -1,0 +1,12 @@
+import PackageDescription
+
+let package = Package(
+    name: "MyCustomPathPackage",
+    targets: [
+        .target(
+            name: "LibraryTarget",
+            dependencies: ["Dep1"],
+            path: "/Sources/MyCustomPath"
+        ),
+    ]
+)
