+++ Tests/SwiftDependencyAuditTests/Fixtures/LargePackage.swift
@@ -1,0 +1,8 @@
+import PackageDescription
+
+let package = Package(
+    name: "LargePackage",
+    targets: [
+        .target(name: "LargeTarget", dependencies: [])
+    ]
+)
