+++ Tests/SwiftDependencyAuditTests/Fixtures/MultiTargetPackage.swift
@@ -1,0 +1,10 @@
+import PackageDescription
+
+let package = Package(
+    name: "MultiTargetPackage",
+    targets: [
+        .target(name: "LibraryTarget", dependencies: ["Dep1"]),
+        .executableTarget(name: "ExecutableTarget", dependencies: ["LibraryTarget"]),
+        .testTarget(name: "TestTarget", dependencies: ["LibraryTarget"])
+    ]
+)
