+++ Tests/SwiftDependencyAuditTests/Fixtures/SimplePackage.swift
@@ -1,0 +1,8 @@
+import PackageDescription
+
+let package = Package(
+    name: "SimplePackage",
+    targets: [
+        .target(name: "SimpleTarget", dependencies: [])
+    ]
+)
