+++ Tests/SwiftDependencyAuditTests/Fixtures/SpecialNamePackage.swift
@@ -1,0 +1,6 @@
+import PackageDescription
+
+let package = Package(
+    name: "My-Special_Package123",
+    targets: []
+)
